[Xtensa] Pass pointers as-is in the Xtensa ABI
Previously, pointers were cast to i32 integers when passed as arguments
to a function or when they are returned. However, this appears to break
the assumption that malloclike functions (such as malloc with the right
attribute) return a pointer, not an integer.

This patch makes sure plain pointers are not coerced in any way and
instead are passed simply as pointers.